requirements.txt
aiohttp==3.10.8
asyncpg==0.29.0
SQLAlchemy==2.0.35
more-itertools==10.5.0
python-dotenv==1.0.1
